lunaa: Set required fp property for anc fingerprint
This fixes the issue with jiiov/anc sensors where during first unlock it complains of fingerprint hw not available and deletes fp data and sets this property, but the subsequent enrolls work fine because the property is set during first unlock attempt. This also requires us not to address the denial to set this property by the oplus fingerprint hal since the hal wants to reset it triggering the bug.

Change-Id: Ifca39bd097ec30613c4bd14c746c609ce562d832
Signed-off-by: dodyirawan85 <40514988+dodyirawan85@users.noreply.github.com>